Move ConnectionPersistence into its own package.
