+++ src/main/java/com/hivemq/mqtt/handler/disconnect/DisconnectHandler.java
@@ -29,8 +29,8 @@ import com.hivemq.limitation.TopicAliasLimiter;
 import com.hivemq.logging.EventLog;
 import com.hivemq.metrics.MetricsHolder;
 import com.hivemq.mqtt.message.disconnect.DISCONNECT;
-import com.hivemq.persistence.ConnectionPersistence;
 import com.hivemq.persistence.clientsession.ClientSessionPersistence;
+import com.hivemq.persistence.connection.ConnectionPersistence;
 import com.hivemq.persistence.util.FutureUtils;
 import com.hivemq.util.ChannelAttributes;
 import com.hivemq.util.Checkpoints;
