+++ src/main/java/com/hivemq/persistence/clientsession/ClientSessionSubscriptionPersistenceImpl.java
@@ -36,11 +36,11 @@ import com.hivemq.mqtt.topic.SubscriptionFlags;
 import com.hivemq.mqtt.topic.TopicFilter;
 import com.hivemq.mqtt.topic.tree.LocalTopicTree;
 import com.hivemq.persistence.AbstractPersistence;
-import com.hivemq.persistence.ConnectionPersistence;
 import com.hivemq.persistence.ProducerQueues;
 import com.hivemq.persistence.SingleWriterService;
 import com.hivemq.persistence.clientsession.SharedSubscriptionServiceImpl.SharedSubscription;
 import com.hivemq.persistence.clientsession.callback.SubscriptionResult;
+import com.hivemq.persistence.connection.ConnectionPersistence;
 import com.hivemq.persistence.local.ClientSessionLocalPersistence;
 import com.hivemq.persistence.local.ClientSessionSubscriptionLocalPersistence;
 import com.hivemq.util.ReasonStrings;
