+++ src/main/java/com/hivemq/persistence/connection/ConnectionPersistence.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.persistence;
+package com.hivemq.persistence.connection;
 
 import com.google.common.util.concurrent.ListenableFuture;
 import com.hivemq.bootstrap.ClientConnection;
