+++ src/test/java/com/hivemq/mqtt/handler/disconnect/DisconnectHandlerTest.java
@@ -27,8 +27,8 @@ import com.hivemq.mqtt.message.ProtocolVersion;
 import com.hivemq.mqtt.message.disconnect.DISCONNECT;
 import com.hivemq.mqtt.message.mqtt5.Mqtt5UserProperties;
 import com.hivemq.mqtt.message.reason.Mqtt5DisconnectReasonCode;
-import com.hivemq.persistence.ConnectionPersistence;
 import com.hivemq.persistence.clientsession.ClientSessionPersistence;
+import com.hivemq.persistence.connection.ConnectionPersistence;
 import com.hivemq.security.auth.ClientData;
 import com.hivemq.util.ChannelAttributes;
 import com.hivemq.util.ChannelUtils;
