+++ src/test/java/com/hivemq/persistence/clientsession/ClientSessionPersistenceImplTest.java
@@ -27,9 +27,9 @@ import com.hivemq.mqtt.message.ProtocolVersion;
 import com.hivemq.mqtt.message.QoS;
 import com.hivemq.mqtt.message.connect.MqttWillPublish;
 import com.hivemq.mqtt.message.reason.Mqtt5DisconnectReasonCode;
-import com.hivemq.persistence.ConnectionPersistenceImpl;
 import com.hivemq.persistence.SingleWriterService;
 import com.hivemq.persistence.clientqueue.ClientQueuePersistence;
+import com.hivemq.persistence.connection.ConnectionPersistenceImpl;
 import com.hivemq.persistence.local.ClientSessionLocalPersistence;
 import com.hivemq.util.ChannelAttributes;
 import io.netty.channel.Channel;
